app/requirements.txt
Django
djangorestframework
requests
mysqlclient
pandas
